Glocal_warriors2/cansat-dashboard/UI/requirements..txt
pyserial
matplotlib